Added status text and clean shutdown to text console
