+++ webterm/text-console-test.js
@@ -2,23 +2,19 @@
 **
 **  Copyright (c) 2023, Phil Claridge
 **
-**  console-text.js
+**  text-console.js
 **    Provides a text mode version of the Cyber Console for CLI use
 **
 **--------------------------------------------------------------------------
 */
 
 /*
-**  console-text.js
+**  text-console.js
 **    Text mode console for DtCyber using the webterm infrastructure
  */
 
-const CyberConsoleText = require('./textconsole/js/cyber-console-text')
-
-const BasicTerminalKitConsole = require('./textconsole/js/basic-terminal-kit')
-
+const CyberConsoleText = require('./textconsole/js/console-text')
 const Machine = require('./textconsole/js/machine-tcp')
-
 const fs = require('fs');
 
 function extractWebSocketPackets(harFilePath) {
@@ -55,14 +51,13 @@ function extractWebsocketMessages(webSocketPackets) {
 }
 
 
-function main_test() {
+function main() {
     // const harFilePath = 'testdata/nos287_running.har';
     const harFilePath = 'testdata/nos287.har';
     const webSocketPackets = extractWebSocketPackets(harFilePath);
     let receivedMessagesBase64 = extractWebsocketMessages(webSocketPackets);
 
-    let basicTerminalKitConsole = new BasicTerminalKitConsole()
-    let consoleText = new CyberConsoleText(basicTerminalKitConsole)
+    let consoleText = new CyberConsoleText()
     consoleText.createScreen();
 
     let index = 0;
@@ -73,7 +68,7 @@ function main_test() {
         if (index >= receivedMessagesBase64.length) {
             index = 0;
         }
-        consoleText.clearScreenBuffer();
+        consoleText.clearScreen();
         let receivedMessagesBase64Element = receivedMessagesBase64[index];
         if (lastReceivedMessagesBase64 !== receivedMessagesBase64Element) {
             const buffer = Buffer.from(receivedMessagesBase64Element, 'base64');
@@ -86,57 +81,6 @@ function main_test() {
         }
 
     }, 200);
-}
-
-
-function main() {
-    let url = 'localhost';
-    let port = 16612;
-    let refreshInterval = 20; // 1/100 sec
-
-    let basicTerminalKitConsole = new BasicTerminalKitConsole()
-    let cyberConsole = new CyberConsoleText(basicTerminalKitConsole)
-    cyberConsole.createScreen();
-
-    const machineId = "nos287";
-    const machine = new Machine(machineId, url, port);
-    machine.setTerminal(cyberConsole);
-    machine.setReceivedDataHandler(data => {
-        cyberConsole.renderText(data);
-    });
-
-    machine.setConnectListener(() => {
-        machine.send(new Uint8Array([0x80, refreshInterval, 0x81]));
-    });
-
-    machine.createConnection();
-
-    // const uplineDataSender = data => {
-    //     machine.send(data);
-    // };
-    // cyberConsole.setUplineDataSender(uplineDataSender);
-    //
-    // machine.setDisconnectListener(() => {
-    //     cyberConsole.displayNotification(1, 128, 128, `Disconnected from ${title}.\n\n   Press any key to reconnect ...`);
-    //     cyberConsole.setUplineDataSender(data => {
-    //         cyberConsole.reset();
-    //         url = machine.createConnection();
-    //         cyberConsole.setUplineDataSender(uplineDataSender);
-    //     });
-    // });
-    // $(document).ajaxError((event, jqxhr, settings, thrownError) => {
-    //     if (settings.url === url) {
-    //         cyberConsole.displayNotification(1, 128, 128, `${jqxhr.responseText}\n\n   Failed to connect to ${title}\n\n   Press any key to try again ...`);
-    //         cyberConsole.setUplineDataSender(data => {
-    //             cyberConsole.reset();
-    //             url = machine.createConnection();
-    //             cyberConsole.setUplineDataSender(uplineDataSender);
-    //         });
-    //     }
-    // });
-    // $(window).bind("beforeunload", () => {
-    //     machine.closeConnection();
-    // });
 }
 
 main();
